refactor(frontend): extract fn to add exercises to workout
